+++ src/Navigation/Menu/Items/Item/Item.js
@@ -1,15 +1,28 @@
-import React from 'react';
+import React, { Component } from 'react';
 import Types from 'prop-types';
 
 import Dropdown from './Dropdown';
+import focusWithin from './focusWithin';
 
-const Item = ({ translationKey, link, items }) => (
-  <li>
-    <a href={link}>{translationKey}</a>
+class Item extends Component {
+  componentDidMount() {
+    focusWithin(this.itemDOMElement);
+  }
 
-    {items && items.length > 0 && <Dropdown items={items} />}
-  </li>
-);
+  render() {
+    return (
+      <li
+        ref={element => {
+          this.itemDOMElement = element;
+        }}
+      >
+        <a href={this.props.link}>{this.props.translationKey}</a>
+
+        {this.props.items && this.props.items.length > 0 && <Dropdown items={this.props.items} />}
+      </li>
+    );
+  }
+}
 
 Item.propTypes = {
   translationKey: Types.string.isRequired,