+++ src/Navigation/Menu/Items/Item/focusWithin/index.js
@@ -1,0 +1,3 @@
+import focusWithin from './focusWithin';
+
+export default focusWithin;